drop i64 from GID, refactor NodeRef, remove internal module
