+++ python/tests/test_graphdb.py
@@ -2365,6 +2365,7 @@ def test_leading_zeroes_ids():
     g.add_node(0, 0)
     g.add_node(1, 0)
     assert g.node(0).history() == [0, 1]
+    assert g.node("0").history() == [0, 1]
     assert g.nodes.name.collect() == ["0"]
