+++ raphtory-api/src/core/entities/mod.rs
@@ -3,6 +3,7 @@ use std::{
     fmt::{Display, Formatter},
 };
 
+use num_traits::ToPrimitive;
 #[cfg(feature = "python")]
 use pyo3::exceptions::PyTypeError;
 #[cfg(feature = "python")]
@@ -10,7 +11,6 @@ use pyo3::prelude::*;
 use serde::{Deserialize, Serialize};
 
 use self::edges::edge_ref::EdgeRef;
-use num_traits::cast::ToPrimitive;
 
 use super::input::input_node::parse_u64_strict;
 
@@ -101,7 +101,6 @@ impl EID {
 #[derive(Debug, Clone, PartialEq, PartialOrd, Eq, Ord, Hash, Serialize, Deserialize)]
 pub enum GID {
     U64(u64),
-    I64(i64),
     Str(String),
 }
 
@@ -115,7 +114,6 @@ impl Display for GID {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
         match self {
             GID::U64(v) => write!(f, "{}", v),
-            GID::I64(v) => write!(f, "{}", v),
             GID::Str(v) => write!(f, "{}", v),
         }
     }
@@ -126,7 +124,6 @@ impl IntoPy<PyObject> for GID {
     fn into_py(self, py: Python<'_>) -> PyObject {
         match self {
             GID::U64(v) => v.into_py(py),
-            GID::I64(v) => v.into_py(py),
             GID::Str(v) => v.into_py(py),
         }
     }
@@ -137,7 +134,6 @@ impl ToPyObject for GID {
     fn to_object(&self, py: Python<'_>) -> PyObject {
         match self {
             GID::U64(v) => v.to_object(py),
-            GID::I64(v) => v.to_object(py),
             GID::Str(v) => v.to_object(py),
         }
     }
@@ -148,11 +144,7 @@ impl<'source> FromPyObject<'source> for GID {
     fn extract(id: &'source PyAny) -> PyResult<Self> {
         id.extract::<String>()
             .map(GID::Str)
-            .or_else(|_| {
-                id.extract::<u64>()
-                    .map(GID::U64)
-                    .or_else(|_| id.extract::<i64>().map(GID::I64))
-            })
+            .or_else(|_| id.extract::<u64>().map(GID::U64))
             .map_err(|_| {
                 let msg = "IDs need to be strings or an unsigned integers";
                 PyTypeError::new_err(msg)
@@ -164,13 +156,6 @@ impl GID {
     pub fn into_str(self) -> Option<String> {
         match self {
             GID::Str(v) => Some(v),
-            _ => None,
-        }
-    }
-
-    pub fn into_i64(self) -> Option<i64> {
-        match self {
-            GID::I64(v) => Some(v),
             _ => None,
         }
     }
@@ -189,13 +174,6 @@ impl GID {
         }
     }
 
-    pub fn as_i64(&self) -> Option<i64> {
-        match self {
-            GID::I64(v) => Some(*v),
-            _ => None,
-        }
-    }
-
     pub fn as_u64(&self) -> Option<u64> {
         match self {
             GID::U64(v) => Some(*v),
@@ -206,7 +184,6 @@ impl GID {
     pub fn to_str(&self) -> Cow<String> {
         match self {
             GID::U64(v) => Cow::Owned(v.to_string()),
-            GID::I64(v) => Cow::Owned(v.to_string()),
             GID::Str(v) => Cow::Borrowed(v),
         }
     }
@@ -214,7 +191,6 @@ impl GID {
     pub fn to_i64(&self) -> Option<i64> {
         match self {
             GID::U64(v) => v.to_i64(),
-            GID::I64(v) => Some(*v),
             GID::Str(v) => parse_u64_strict(v)?.to_i64(),
         }
     }
@@ -222,7 +198,6 @@ impl GID {
     pub fn to_u64(&self) -> Option<u64> {
         match self {
             GID::U64(v) => Some(*v),
-            GID::I64(v) => v.to_u64(),
             GID::Str(v) => parse_u64_strict(v),
         }
     }
@@ -231,12 +206,6 @@ impl GID {
 impl From<u64> for GID {
     fn from(id: u64) -> Self {
         Self::U64(id)
-    }
-}
-
-impl From<i64> for GID {
-    fn from(id: i64) -> Self {
-        Self::I64(id)
     }
 }
 
@@ -256,7 +225,6 @@ impl<'a> From<GidRef<'a>> for GID {
     fn from(value: GidRef<'a>) -> Self {
         match value {
             GidRef::U64(v) => GID::U64(v),
-            GidRef::I64(v) => GID::I64(v),
             GidRef::Str(v) => GID::Str(v.to_owned()),
         }
     }
@@ -265,15 +233,22 @@ impl<'a> From<GidRef<'a>> for GID {
 #[derive(Copy, Clone, Debug, PartialEq, PartialOrd, Eq, Ord, Hash)]
 pub enum GidRef<'a> {
     U64(u64),
-    I64(i64),
     Str(&'a str),
+}
+
+impl Display for GidRef<'_> {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        match self {
+            GidRef::U64(v) => write!(f, "{}", v),
+            GidRef::Str(v) => write!(f, "{}", v),
+        }
+    }
 }
 
 impl<'a> From<&'a GID> for GidRef<'a> {
     fn from(value: &'a GID) -> Self {
         match value {
             GID::U64(v) => GidRef::U64(*v),
-            GID::I64(v) => GidRef::I64(*v),
             GID::Str(v) => GidRef::Str(v),
         }
     }
@@ -283,13 +258,6 @@ impl<'a> GidRef<'a> {
     pub fn as_str(self) -> Option<&'a str> {
         match self {
             GidRef::Str(s) => Some(s),
-            _ => None,
-        }
-    }
-
-    pub fn as_i64(self) -> Option<i64> {
-        match self {
-            GidRef::I64(v) => Some(v),
             _ => None,
         }
     }
@@ -304,7 +272,6 @@ impl<'a> GidRef<'a> {
     pub fn to_owned(self) -> GID {
         match self {
             GidRef::U64(v) => GID::U64(v),
-            GidRef::I64(v) => GID::I64(v),
             GidRef::Str(v) => GID::Str(v.to_owned()),
         }
     }
@@ -312,7 +279,6 @@ impl<'a> GidRef<'a> {
     pub fn to_str(self) -> Cow<'a, str> {
         match self {
             GidRef::U64(v) => Cow::Owned(v.to_string()),
-            GidRef::I64(v) => Cow::Owned(v.to_string()),
             GidRef::Str(v) => Cow::Borrowed(v),
         }
     }
@@ -320,7 +286,6 @@ impl<'a> GidRef<'a> {
     pub fn to_i64(self) -> Option<i64> {
         match self {
             GidRef::U64(v) => v.to_i64(),
-            GidRef::I64(v) => Some(v),
             GidRef::Str(v) => parse_u64_strict(v)?.to_i64(),
         }
     }
@@ -328,7 +293,6 @@ impl<'a> GidRef<'a> {
     pub fn to_u64(self) -> Option<u64> {
         match self {
             GidRef::U64(v) => Some(v),
-            GidRef::I64(v) => v.to_u64(),
             GidRef::Str(v) => parse_u64_strict(v),
         }
     }
