+++ raphtory-benchmark/benches/parameterized.rs
@@ -2,6 +2,7 @@ use crate::common::{bootstrap_graph, run_large_ingestion_benchmarks};
 use criterion::{
     criterion_group, criterion_main, AxisScale, Criterion, PlotConfiguration, Throughput,
 };
+use raphtory_api::core::entities::GID;
 
 mod common;
 
@@ -12,11 +13,17 @@ pub fn parameterized(c: &mut Criterion) {
     let mut ingestion_group = c.benchmark_group("ingestion-num_nodes");
     ingestion_group.plot_config(PlotConfiguration::default().summary_scale(AxisScale::Logarithmic));
     for num_nodes in nodes {
-        let make_graph = || bootstrap_graph(num_nodes);
+        let make_graph = || bootstrap_graph(num_nodes, |id| GID::U64(id));
+        let make_graph_str = || bootstrap_graph(num_nodes, |id| GID::Str(id.to_string()));
         ingestion_group.throughput(Throughput::Elements(num_nodes as u64));
         ingestion_group.sample_size(10);
         ingestion_group.warm_up_time(std::time::Duration::from_secs(1));
-        run_large_ingestion_benchmarks(&mut ingestion_group, make_graph, Some(num_nodes));
+        run_large_ingestion_benchmarks(
+            &mut ingestion_group,
+            make_graph,
+            make_graph_str,
+            Some(num_nodes),
+        );
     }
     ingestion_group.finish();
 }
