+++ raphtory/src/core/entities/nodes/node_ref.rs
@@ -5,8 +5,7 @@ use raphtory_api::core::entities::{GidRef, GID};
 #[derive(Copy, Clone, PartialOrd, PartialEq, Debug)]
 pub enum NodeRef<'a> {
     Internal(VID),
-    External(u64),
-    ExternalStr(&'a str),
+    External(GidRef<'a>),
 }
 
 pub trait AsNodeRef {
@@ -18,16 +17,14 @@ pub trait AsNodeRef {
     {
         match self.as_node_ref() {
             NodeRef::Internal(vid) => Either::Right(vid),
-            NodeRef::External(u) => Either::Left(GID::U64(u)),
-            NodeRef::ExternalStr(s) => Either::Left(GID::Str(s.to_string())),
+            NodeRef::External(gid) => Either::Left(gid.into()),
         }
     }
 
     fn as_gid_ref(&self) -> Either<GidRef, VID> {
         match self.as_node_ref() {
             NodeRef::Internal(vid) => Either::Right(vid),
-            NodeRef::External(u) => Either::Left(GidRef::U64(u)),
-            NodeRef::ExternalStr(s) => Either::Left(GidRef::Str(s)),
+            NodeRef::External(u) => Either::Left(u),
         }
     }
 }
@@ -46,19 +43,19 @@ impl AsNodeRef for VID {
 
 impl AsNodeRef for u64 {
     fn as_node_ref(&self) -> NodeRef {
-        NodeRef::External(*self)
+        NodeRef::External(GidRef::U64(*self))
     }
 }
 
 impl AsNodeRef for String {
     fn as_node_ref(&self) -> NodeRef {
-        NodeRef::ExternalStr(self.as_ref())
+        NodeRef::External(GidRef::Str(&self))
     }
 }
 
 impl<'a> AsNodeRef for &'a str {
     fn as_node_ref(&self) -> NodeRef {
-        NodeRef::ExternalStr(self)
+        NodeRef::External(GidRef::Str(self))
     }
 }
 
@@ -70,11 +67,8 @@ impl<'a, V: AsNodeRef> AsNodeRef for &'a V {
 
 impl AsNodeRef for GID {
     fn as_node_ref(&self) -> NodeRef {
-        match self {
-            GID::U64(u) => NodeRef::External(*u),
-            GID::Str(s) => NodeRef::ExternalStr(s),
-            GID::I64(i) => NodeRef::External(*i as u64),
-        }
+        let gid_ref: GidRef = self.into();
+        NodeRef::External(gid_ref)
     }
 }
