--- raphtory/src/db/internal/list_ops.rs
@@ -1,18 +1,0 @@
-use crate::{
-    core::entities::graph::tgraph::InternalGraph,
-    db::api::view::internal::{EdgeList, ListOps, NodeList},
-};
-
-impl ListOps for InternalGraph {
-    fn node_list(&self) -> NodeList {
-        NodeList::All {
-            num_nodes: self.inner().storage.nodes_len(),
-        }
-    }
-
-    fn edge_list(&self) -> EdgeList {
-        EdgeList::All {
-            num_edges: self.inner().storage.edges_len(),
-        }
-    }
-}
