--- raphtory/src/db/internal/mod.rs
@@ -1,12 +1,0 @@
-pub(crate) mod addition;
-pub(crate) mod core_ops;
-pub(crate) mod deletion;
-pub(crate) mod edge_filter_ops;
-pub(crate) mod layer_ops;
-mod list_ops;
-pub(crate) mod materialize;
-mod node_filter_ops;
-pub(crate) mod prop_add;
-pub(crate) mod static_properties;
-pub(crate) mod temporal_properties;
-pub(crate) mod time_semantics;
