+++ raphtory/src/graph.proto
@@ -77,7 +77,6 @@ message GraphMeta{
 message GID {
   oneof gid {
     uint64 gid_u64 = 1;
-    int64 gid_i64 = 2;
     string  gid_str = 3;
   }
 }
@@ -85,7 +84,6 @@ message GID {
 message Node{
   GID gid = 1;
   uint64 vid = 2;
-  optional string name = 3;
 }
 
 message PropertiesMeta{
