+++ raphtory/src/graphgen/mod.rs
@@ -12,7 +12,6 @@ pub(crate) fn next_id<'graph, G: GraphViewOps<'graph>>(g: &G, max_gid: Option<GI
     let max_gid = max_gid.unwrap_or_else(|| g.nodes().id().max().unwrap_or(GID::U64(0)));
     match max_gid {
         GID::U64(id) => GID::U64(id + 1),
-        GID::I64(id) => GID::I64(id + 1),
         GID::Str(_) => {
             let mut rng = rand::thread_rng();
             loop {
